+++ changelog.md
@@ -7,6 +7,7 @@
 - Layers!
 	- You can make stuff on a layer directly, or shift items from other layers via a dropdown in their details popup.
 	- Relationship lines can be created from and to characters on separate layers, and be on different layers from the things they connect
+- Can now export image in WEBP and JPEG formats as well as PNG; these extra formats have a slider for image quality.
 # v0.7.0
 - New app icon.
 - Added github repo link to Settings > About.